change routes, add team info page
